# go/model_beneficiary.go
/*
 * Beneficiary API
 *
 * This is a simple Beneficiary API, where you can Create, Read, and Delete beneficiary
 *
 * API version: 1.0.0
 * Contact: mjangid@cynergybank.co.uk
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger
import (
	"time"
)

type Beneficiary struct {

	BeneficiaryId int32 `json:"beneficiary_id,omitempty"`

	BeneficiaryName string `json:"beneficiary_name"`

	BeneficiaryScheme string `json:"beneficiary_scheme"`

	BeneficiarySchemeId string `json:"beneficiary_scheme_id"`

	BeneficiaryAddress *interface{} `json:"beneficiary_address"`

	BeneficiarySecondaryIdentifier string `json:"beneficiary_secondary_identifier,omitempty"`

	BeneficiaryAvatar *interface{} `json:"beneficiary_avatar,omitempty"`

	PartyId string `json:"party_id"`

	Metadata *interface{} `json:"metadata,omitempty"`

	DateCreated time.Time `json:"date_created,omitempty"`
}